create layout list restaurant

+++ lib/data/restaurant.dart
@@ -1,4 +1,4 @@
-import 'menu.dart';
+import 'menus.dart';
 
 class Restaurant {
   String? id;
@@ -24,7 +24,7 @@ class Restaurant {
     description = json["description"];
     pictureId = json["pictureId"];
     city = json["city"];
-    rating = json["rating"];
+    rating = json["rating"].toDouble();
     menus = json["menus"] == null ? null : Menus.fromJson(json["menus"]);
   }
